backend/requirements.txt
fastapi
uvicorn
python-multipart

sqlalchemy>=2.0
alembic>=1.13
python-dotenv>=1.0

python-jose[cryptography]
passlib[bcrypt]
pydantic[email]

torch
torchvision
transformers
pillow
numpy
scipy
